+++ config/console.php
@@ -1,11 +1,12 @@
 <?php
 
 return array (
-  'commands' => 
+  'commands' =>
   array (
-    'task' => 'app\\common\\command\\Task',
-    'we' => 'app\\common\\command\\WeCenter',
-    'sitemap:build' => 'app\\common\\command\\Sitemap',
-    'insight:report' => 'app\\common\\command\\InsightReport',
-  ),
+        'task' => 'app\\common\\command\\Task',
+        'we' => 'app\\common\\command\\WeCenter',
+        'sitemap:build' => 'app\\common\\command\\Sitemap',
+        'insight:report' => 'app\\common\\command\\InsightReport',
+        'api:doc' => 'app\\common\\command\\ApiDoc',
+    ),
 );